rename Snackbar.tsx to index.tsx
