+++ src/snackbar/tests/unit/Snackbar.tsx
@@ -5,7 +5,7 @@ import * as sinon from 'sinon';
 import assertationTemplate from '@dojo/framework/testing/assertionTemplate';
 import harness from '@dojo/framework/testing/harness';
 import { tsx } from '@dojo/framework/widget-core/tsx';
-import Snackbar from '../../Snackbar';
+import Snackbar from '../../index';
 import * as css from '../../../theme/snackbar.m.css';
 import Button from '../../../button';
 import bundle from '../../nls/Snackbar';
